chi graph tmp ok`

+++ examples/conversion.h
@@ -19,7 +19,7 @@
     (((double)(tval).tv_sec) + (((double)(tval).tv_usec) / 1000000.0))
 
 // Converts a byte (integral) value to mebibytes (floating-point)
-#define BYTE_TO_MIB(size) (((double)(size)) / (1024.0 * 1024.0))
+#define BYTE_TO_MIB(size) (((long)(size)) / (1024 * 1024))
 
 // Converts a mebibyte (floating-point) value to bytes (integral)
 #define MIB_TO_BYTE(size) ((size_t)((size) * 1024.0 * 1024.0))
